Updated tools

+++ src/start.c
@@ -106,9 +106,9 @@ const DeviceVectors exception_table = {
   /* Configure Initial Stack Pointer, using linker-generated symbols */
   (void*)(&_estack),
 
-  (void*)Reset_Handler,
-  (void*)NMI_Handler,
-  (void*)HardFault_Handler,
+  (void*)(uintptr_t)Reset_Handler,
+  (void*)(uintptr_t)NMI_Handler,
+  (void*)(uintptr_t)HardFault_Handler,
   (void*)(0UL), /* Reserved */
   (void*)(0UL), /* Reserved */
   (void*)(0UL), /* Reserved */
@@ -116,41 +116,41 @@ const DeviceVectors exception_table = {
   (void*)(0UL), /* Reserved */
   (void*)(0UL), /* Reserved */
   (void*)(0UL), /* Reserved */
-  (void*)SVC_Handler,
+  (void*)(uintptr_t)SVC_Handler,
   (void*)(0UL), /* Reserved */
   (void*)(0UL), /* Reserved */
-  (void*)PendSV_Handler,
-  (void*)SysTick_Handler,
+  (void*)(uintptr_t)PendSV_Handler,
+  (void*)(uintptr_t)SysTick_Handler,
 
   /* Configurable interrupts */
-  (void*)PM_Handler, /*  0 Power Manager */
-  (void*)SYSCTRL_Handler, /*  1 System Control */
-  (void*)WDT_Handler, /*  2 Watchdog Timer */
-  (void*)RTC_Handler, /*  3 Real-Time Counter */
-  (void*)EIC_Handler, /*  4 External Interrupt Controller */
-  (void*)NVMCTRL_Handler, /*  5 Non-Volatile Memory Controller */
-  (void*)DMAC_Handler, /*  6 Direct Memory Access Controller */
-  (void*)USB_Handler, /*  7 Universal Serial Bus */
-  (void*)EVSYS_Handler, /*  8 Event System Interface */
-  (void*)SERCOM0_Handler, /*  9 Serial Communication Interface 0 */
-  (void*)SERCOM1_Handler, /* 10 Serial Communication Interface 1 */
-  (void*)SERCOM2_Handler, /* 11 Serial Communication Interface 2 */
-  (void*)SERCOM3_Handler, /* 12 Serial Communication Interface 3 */
-  (void*)SERCOM4_Handler, /* 13 Serial Communication Interface 4 */
-  (void*)SERCOM5_Handler, /* 14 Serial Communication Interface 5 */
-  (void*)TCC0_Handler, /* 15 Timer Counter Control 0 */
-  (void*)TCC1_Handler, /* 16 Timer Counter Control 1 */
-  (void*)TCC2_Handler, /* 17 Timer Counter Control 2 */
-  (void*)TC3_Handler, /* 18 Basic Timer Counter 0 */
-  (void*)TC4_Handler, /* 19 Basic Timer Counter 1 */
-  (void*)TC5_Handler, /* 20 Basic Timer Counter 2 */
-  (void*)TC6_Handler, /* 21 Basic Timer Counter 3 */
-  (void*)TC7_Handler, /* 22 Basic Timer Counter 4 */
-  (void*)ADC_Handler, /* 23 Analog Digital Converter */
-  (void*)AC_Handler, /* 24 Analog Comparators */
-  (void*)DAC_Handler, /* 25 Digital Analog Converter */
-  (void*)PTC_Handler, /* 26 Peripheral Touch Controller */
-  (void*)I2S_Handler, /* 27 Inter-IC Sound Interface */
+  (void*)(uintptr_t)PM_Handler, /*  0 Power Manager */
+  (void*)(uintptr_t)SYSCTRL_Handler, /*  1 System Control */
+  (void*)(uintptr_t)WDT_Handler, /*  2 Watchdog Timer */
+  (void*)(uintptr_t)RTC_Handler, /*  3 Real-Time Counter */
+  (void*)(uintptr_t)EIC_Handler, /*  4 External Interrupt Controller */
+  (void*)(uintptr_t)NVMCTRL_Handler, /*  5 Non-Volatile Memory Controller */
+  (void*)(uintptr_t)DMAC_Handler, /*  6 Direct Memory Access Controller */
+  (void*)(uintptr_t)USB_Handler, /*  7 Universal Serial Bus */
+  (void*)(uintptr_t)EVSYS_Handler, /*  8 Event System Interface */
+  (void*)(uintptr_t)SERCOM0_Handler, /*  9 Serial Communication Interface 0 */
+  (void*)(uintptr_t)SERCOM1_Handler, /* 10 Serial Communication Interface 1 */
+  (void*)(uintptr_t)SERCOM2_Handler, /* 11 Serial Communication Interface 2 */
+  (void*)(uintptr_t)SERCOM3_Handler, /* 12 Serial Communication Interface 3 */
+  (void*)(uintptr_t)SERCOM4_Handler, /* 13 Serial Communication Interface 4 */
+  (void*)(uintptr_t)SERCOM5_Handler, /* 14 Serial Communication Interface 5 */
+  (void*)(uintptr_t)TCC0_Handler, /* 15 Timer Counter Control 0 */
+  (void*)(uintptr_t)TCC1_Handler, /* 16 Timer Counter Control 1 */
+  (void*)(uintptr_t)TCC2_Handler, /* 17 Timer Counter Control 2 */
+  (void*)(uintptr_t)TC3_Handler, /* 18 Basic Timer Counter 0 */
+  (void*)(uintptr_t)TC4_Handler, /* 19 Basic Timer Counter 1 */
+  (void*)(uintptr_t)TC5_Handler, /* 20 Basic Timer Counter 2 */
+  (void*)(uintptr_t)TC6_Handler, /* 21 Basic Timer Counter 3 */
+  (void*)(uintptr_t)TC7_Handler, /* 22 Basic Timer Counter 4 */
+  (void*)(uintptr_t)ADC_Handler, /* 23 Analog Digital Converter */
+  (void*)(uintptr_t)AC_Handler, /* 24 Analog Comparators */
+  (void*)(uintptr_t)DAC_Handler, /* 25 Digital Analog Converter */
+  (void*)(uintptr_t)PTC_Handler, /* 26 Peripheral Touch Controller */
+  (void*)(uintptr_t)I2S_Handler, /* 27 Inter-IC Sound Interface */
   (void*)(0UL) /* Reserved */
 };
